@@ -5,7 +5,7 @@
   "main": "server.js",
   "type": "module",
   "engines": {
-    "node": "12.11.1"
+    "node": "14.15.1"
   },
   "scripts": {
     "start": "node backend/server",
