@@ -1,18 +1,18 @@
-const mongoose = require('mongoose')
+import mongoose from "mongoose";
 
 const connectDB = async () => {
-	try {
-		const conn = await mongoose.connect(process.env.MONGO_URL, {
-			useUnifiedTopology: true,
-			useNewUrlParser: true,
-			useCreateIndex: true,
-		})
+  try {
+    const conn = await mongoose.connect(process.env.MONGO_URL, {
+      useUnifiedTopology: true,
+      useNewUrlParser: true,
+      useCreateIndex: true,
+    });
 
-		console.log(`MongoDB Connected: ${conn.connection.host}`.cyan.underline)
-	} catch (error) {
-		console.error(`Error: ${error.message}`.red.underline.bold)
-		process.exit(1)
-	}
-}
+    console.log(`MongoDB Connected: ${conn.connection.host}`.cyan.underline);
+  } catch (error) {
+    console.error(`Error: ${error.message}`.red.underline.bold);
+    process.exit(1);
+  }
+};
 
-export default connectDB
+export default connectDB;
