@@ -1,57 +1,57 @@
-const mongoose = require('mongoose')
-const dotenv = require('dotenv')
-const colors = require('colors')
-const users = require('./data/users.js')
-const products = require('./data/products.js')
-const User = require('./models/userModel.js')
-const Product = require('./models/productModel.js')
-const Order = require('./models/orderModel.js')
-const connectDB = require('./config/db.js')
+import mongoose from 'mongoose'
+import dotenv from 'dotenv'
+import colors from 'colors'
+import users from './data/users.js'
+import products from './data/products.js'
+import User from './models/userModel.js'
+import Product from './models/productModel.js'
+import Order from './models/orderModel.js'
+import connectDB from './config/db.js'
 
 dotenv.config()
 
 connectDB()
 
 const importData = async () => {
-	try {
-		await Order.deleteMany()
-		await Product.deleteMany()
-		await User.deleteMany()
+  try {
+    await Order.deleteMany()
+    await Product.deleteMany()
+    await User.deleteMany()
 
-		const createdUsers = await User.insertMany(users)
+    const createdUsers = await User.insertMany(users)
 
-		const adminUser = createdUsers[0]._id
+    const adminUser = createdUsers[0]._id
 
-		const sampleProducts = products.map((product) => {
-			return { ...product, user: adminUser }
-		})
+    const sampleProducts = products.map((product) => {
+      return { ...product, user: adminUser }
+    })
 
-		await Product.insertMany(sampleProducts)
+    await Product.insertMany(sampleProducts)
 
-		console.log('Data Imported!'.green.inverse)
-		process.exit()
-	} catch (error) {
-		console.error(`${error}`.red.inverse)
-		process.exit(1)
-	}
+    console.log('Data Imported!'.green.inverse)
+    process.exit()
+  } catch (error) {
+    console.error(`${error}`.red.inverse)
+    process.exit(1)
+  }
 }
 
 const destroyData = async () => {
-	try {
-		await Order.deleteMany()
-		await Product.deleteMany()
-		await User.deleteMany()
+  try {
+    await Order.deleteMany()
+    await Product.deleteMany()
+    await User.deleteMany()
 
-		console.log('Data Destroyed!'.red.inverse)
-		process.exit()
-	} catch (error) {
-		console.error(`${error}`.red.inverse)
-		process.exit(1)
-	}
+    console.log('Data Destroyed!'.red.inverse)
+    process.exit()
+  } catch (error) {
+    console.error(`${error}`.red.inverse)
+    process.exit(1)
+  }
 }
 
 if (process.argv[2] === '-d') {
-	destroyData()
+  destroyData()
 } else {
-	importData()
+  importData()
 }
