@@ -1,6 +1,6 @@
-const jwt =require('jsonwebtoken')
-const asyncHandler =require('express-async-handler')
-const User =require('../models/userModel.js')
+import jwt from 'jsonwebtoken'
+import asyncHandler from 'express-async-handler'
+import User from '../models/userModel.js'
 
 const protect = asyncHandler(async (req, res, next) => {
   let token
