@@ -1,5 +1,5 @@
-const mongoose =require('mongoose')
-const bcrypt =require('bcryptjs')
+import mongoose from 'mongoose'
+import bcrypt from 'bcryptjs'
 
 const userSchema = mongoose.Schema(
   {
