@@ -1,14 +1,14 @@
-const express =require('express')
+import express from 'express'
 const router = express.Router()
-const {
+import {
   addOrderItems,
   getOrderById,
   updateOrderToPaid,
   updateOrderToDelivered,
   getMyOrders,
   getOrders,
-} =require('../controllers/orderController.js')
-const { protect, admin } =require('../middleware/authMiddleware.js')
+} from '../controllers/orderController.js'
+import { protect, admin } from '../middleware/authMiddleware.js'
 
 router.route('/').post(protect, addOrderItems).get(protect, admin, getOrders)
 router.route('/myorders').get(protect, getMyOrders)
