@@ -1,6 +1,6 @@
-const express =require('express')
+import express from 'express'
 const router = express.Router()
-const {
+import {
   getProducts,
   getProductById,
   deleteProduct,
@@ -8,8 +8,8 @@ const {
   updateProduct,
   createProductReview,
   getTopProducts,
-} =require('../controllers/productController.js')
-const { protect, admin } =require('../middleware/authMiddleware.js')
+} from '../controllers/productController.js'
+import { protect, admin } from '../middleware/authMiddleware.js'
 
 router.route('/').get(getProducts).post(protect, admin, createProduct)
 router.route('/:id/reviews').post(protect, createProductReview)
